+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.20.4 (unreleased)
+
+* *(client)* Changed media download methods to log requests and to raise
+  exceptions on non-successful status codes.
+
 ## v0.20.3 (2023-11-10)
 
 * *(client)* Deprecated MSC2716 methods and added new Beeper-specific batch